refactor: split linker edit reply tasks
